Updated Locations api

+++ src/main/kotlin/net/axay/kspigot/structures/Circle.kt
@@ -56,17 +56,17 @@ abstract class Circle(val radius: Number) {
 
     final fun buildAtX(loc: Location) {
         for (it in fillLocations)
-            setAt(Location(loc.world, loc.x, loc.y + it.x.toDouble(), loc.z + it.x.toDouble()))
+            setAt(Location(loc.world, loc.x, loc.y + it.x, loc.z + it.x))
     }
 
     final fun buildAtY(loc: Location) {
         for (it in fillLocations)
-            setAt(Location(loc.world, loc.x + it.x.toDouble(), loc.y, loc.z + it.x.toDouble()))
+            setAt(Location(loc.world, loc.x + it.x, loc.y, loc.z + it.x))
     }
 
     final fun buildAtZ(loc: Location) {
         for (it in fillLocations)
-            setAt(Location(loc.world, loc.x + it.x.toDouble(), loc.y + it.x.toDouble(), loc.z))
+            setAt(Location(loc.world, loc.x + it.x, loc.y + it.x, loc.z))
     }
 
     abstract fun setAt(loc: Location)
